adding error handling

+++ src/main/java/com/example/demo/studentController.java
@@ -2,6 +2,9 @@ package com.example.demo;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -17,6 +20,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.example.demo.error.studentNotFound;
+
 @RestController
 @RequestMapping("/students")
 public class studentController {
@@ -25,7 +30,7 @@ public class studentController {
 	 studentRepo repo;
       
       @PostMapping("/post")
-      public ResponseEntity<student> addPerson(student person){
+      public ResponseEntity<student> addPerson(@Valid student person){
           
           student newPerson = repo.save(person);       
           
@@ -54,19 +59,33 @@ public class studentController {
     	// return "reshma";
       }
       @PutMapping("/update/{id}")
-      public String getAllUpdatedData(@PathVariable Integer id,student std){
+      public String getAllUpdatedData(@PathVariable Integer id,student std) throws studentNotFound{
     	  
-    	  Long id1 = Long.valueOf(id);
+			/*
+			 * Long id1 = Long.valueOf(id); student s = repo.findById(id1).get();
+			 * s.setName(std.getName()); repo.save(s);
+			 */
 		
-		 student s = repo.findById(id1).get();
-		 s.setName(std.getName()); 
-		 repo.save(s);
+		 Long id1 = Long.valueOf(id);
+		 Optional<student> s = repo.findById(id1);
+		 if(!s.isPresent()) {
+			 throw new studentNotFound("Department Not Available");
+		 }
+		 
+		 student s1 = s.get();
+		 s1.setName(std.getName()); 
+		 repo.save(s1);
 			 
     	  return "data has been updated";  
       }
 	  @DeleteMapping("/delete/{id}")
-      public String deleteData(@PathVariable Integer id) {
+      public String deleteData(@PathVariable Integer id) throws studentNotFound {
 		  Long id1 = Long.valueOf(id);
+		  Optional<student> s = repo.findById(id1);
+		  if(!s.isPresent()) {
+			  throw new studentNotFound("Id is not available");
+		  }
+
     	  repo.deleteById(id1);
 		  return "data has been deleted";
       }
